feat: sellers ads and related ads query

+++ database/migrations/1691408776375_products.ts
@@ -7,8 +7,8 @@ export default class extends BaseSchema {
     this.schema.createTable(this.tableName, (table) => {
       table.increments('id')
       table.string('title').notNullable()
-      table.decimal('price').notNullable()
-      table.decimal('previous_price').nullable()
+      table.integer('price').notNullable()
+      table.integer('previous_price').nullable()
       table.integer('views').defaultTo(0)
       table.text('description', 'longtext').notNullable()
       table
